Add files via upload
Color Detection of Traffic lights using opencv python

The aim of this project is to detect the color of traffic lights using opencv library in python . Hough transform method is used to detect the different shapes present in an image for computation. This is a bit more efficient than using 3d accumulator matrix method.